Implement compatible authentication system
The login system for demokratia is designed to be compatible with the
main website's, so users can authenticate using their regular
credentials in order to vote.

Some pieces of code (like the UserModel schema) have been taken from the
website's repo, Repo-IEEEsb/Web-IEEEsb. Credits for that go to Carlos
Cuenca (@charlascuen) and Gregorio Juliana (@thunkar).