Implement a global error handler
Add a middleware that is capable of handling any errors thrown in the
application.

A DemokratiaError class has been defined, and any controlled exceptions
should inherit from it (i.e. exceptions we throw for errors contemplated in
the application's logic).

Any error that doesn't inherit from DemokratiaError is considered to be
completely unexpected and produces an HTTP 500 error.