Install mongoose-float via HTTPS
We use our own version of mongoose-float to avoid problems related to
peer dependencies. However, the link in the project's package.json was
for connection via SSH, which requires authentication.

Since that doesn't make sense at all when our fork of mongoose-float is
public, the dependency is now set up for downloading via HTTPS.