Hide longDescription from listElections queries
Showing the longDescription for each Election in the database may result
in a very verbose response, and it isn't a field that shouldn't be shown
on a global level (i.e. when displaying all the Elections).

The longDescription, as well as other fields that are more specific,
should be fetched by querying the specific Election.