Remove "unique" property from Elections' remainingVoters
Even though we don't want duplicate items in remainingVoters, the
"unique" property doesn't work as expected: it keeps the items unique
across ALL "Election" objects, making it useless for our needs.

The uniqueness of the remaining voters array must be guaranteed by the
backend, which for now is fine since we only add items once in a
duplicate-free operation.