Add npm script for automatic Angular dev builds
The script runs in the background "ng build --watch", which
automatically looks for changes in the frontend's files every second and
triggers a live build if required.

"ng build --watch" is used instead of "ng serve" because demokratia's
server is designed to serve Angular's files. Both "ng serve" and backend
servers can't be run at the same time, because they have to be bound to
the same port; hence having "ng build --watch" make the builds
automatically, and let demokratia serve those.