Make "npm run build" hide progress
The progress animation produced by Angular's CLI causes problems
whenever the output is stored to a log (like we do with Jenkins), making
it hard to read the build's output.

It's still possible to know that the build has finished whenever "ng
build" outputs the sizes of the different bundles.