Remove unnecessary populations
For electionController: "Poll" objects are directly stored inside the
"Elections" as subdocuments. And even if they were ObjectId references,
the "Election" objects don't have a "poll" key (it's "polls").

For pollController: the candidate array returned by Election.find is
expected to contain ObjectIds, in order to use them in the $nin part of
the next query. Had this population worked properly, the endpoint would
be broken.

Given this analysis, we can assume these population are useless and safe
to remove.