Add censusSize to the Election schema
There was no way to know the relation of people who voted/people who
could vote, because remainingVoters is designed to change its size as
people cast their votes.

This value should be initialized whenever the "Election" object is
created, to indicate the maximum amount of members that can vote.