Replace express-validation with celebrate
express-validation and celebrate both provide a middleware that wraps
Joi's validator. However, express-validation doesn't allow passing
custom settings to Joi, which was making the "stripUnknown: true"
setting completely useless.

In order to avoid having to specify the options every time the validator
is called, the "validate" and "validateWithoutStripping" functions have
been created, to wrap celebrate's validator with the desired settings
on.

More details on those two functions can be found commented in the
validators file.

Also, the joi package is no longer a dependency because celebrate
provides its own version.