Implement endpoint for voting
An endpoint for voting, with the usual checks one might expect (the
voted user is a valid candidate, the poll exists, etc.). Blank votes can
be cast by setting the "candidate" field to null.

A new model, "Ballot", has been created to store the set of choices made
by a user in an election when voting.

The reason for adding the votes in the "Election" object rather than
each independent poll is mostly for minimizing the number of queries
when reading/writing votes, and being able to easily have a token to let
the users check that their ballots are there.

Soon users will be able to find their ballots using that token (the
"Ballot" object's ID) with a new endpoint.