Implement endpoint for fetching election results
Once an election's voting period has ended (i.e. past the endDate), a
list of the polls with the amount of votes per candidate can be fetched
with this endpoint.

Given that the "Ballot" data structure is a bit more complicated than we
expected, counting the votes requires a more complex query (hence the
long .aggregate call). Fortunately it's a process that doesn't take too
many steps and we can take a user-friendly result straight from the
output.